app/src/services/inference/compute/vlm_server/requirements.txt
# Full dependencies for local development and server
# For Modal deployment, use requirements-modal.txt instead (minimal dependencies)

# Core inference dependencies
pydantic==2.12.2
torch==2.8.0
transformers==4.57.1
Pillow==10.4.0

# Local development only
python-dotenv==1.0.1  # For loading .env files locally

# Model-specific dependencies
google-genai>=1.0.0

# gRPC server dependencies (local server only, not needed in Modal)
protobuf==6.33.0
grpcio-tools==1.76.0
grpcio==1.76.0

# Modal client (for deployment)
modal==1.2.1
